requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
python-multipart

# Database  
sqlalchemy==1.4.53
psycopg2-binary
alembic

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv
PyJWT  # JWT token generation for temporary access

# Validation
pydantic
pydantic[email]

# CORS
fastapi-cors

# Testing
pytest
pytest-asyncio
httpx

# Document Processing
PyMuPDF  # PDF text extraction and redaction (fitz)
PyPDF2  # PDF manipulation
python-docx
pdfplumber
reportlab  # PDF generation and manipulation
openpyxl  # Excel file generation

# Cloud Storage
boto3

# AI/ML and Vector DB
google-genai  # NEW Google Gemini SDK (replaces google-generativeai)
langchain
langchain-google-genai
langchain-huggingface
chromadb
sentence-transformers
transformers
torch
numpy
